Refactor: Cleaning code of ban-core-api
