+++ apps/ban-core-api/src/routes/data/index.ts
@@ -1,10 +1,10 @@
 import 'dotenv/config.js' // eslint-disable-line import/no-unassigned-import
 import express from 'express'
 
-const app = express.Router()
+import banRoutes from './ban/ban.js'
+import diffRoutes from './diff/diff.js'
 
-import banRoutes from './ban.js'
-import diffRoutes from './diff.js'
+const app = express.Router()
 
 app.use('/ban', banRoutes);
 app.use('/diff', diffRoutes);
